+++ Sources/iOSClientExposure/Entitlements/Request/ValidateEntitlement.swift
@@ -21,14 +21,15 @@ public struct ValidateEntitlement: ExposureType {
     /// `SessionToken` identifying the user making the request
     public let sessionToken: SessionToken
     
-    public let  programStartTime: String?
+    /// date/time to run the entitlement for ( ex : programStartTime + 1ms )
+    public let  entitlementDate: String?
     
 
-    public init(assetId: String, environment: Environment, sessionToken: SessionToken, programStartTime: String? = nil  ) {
+    public init(assetId: String, environment: Environment, sessionToken: SessionToken, entitlementDate: String? = nil  ) {
         self.assetId = assetId
         self.environment = environment
         self.sessionToken = sessionToken
-        self.programStartTime = programStartTime
+        self.entitlementDate = entitlementDate
     }
     
     public var endpointUrl: String {
@@ -39,8 +40,8 @@ public struct ValidateEntitlement: ExposureType {
     
     public var parameters: [String: Any] {
         var parameters: [String: String] = [:]
-        if let programStartTime = programStartTime {
-            parameters["time"] = programStartTime
+        if let entitlementDate = entitlementDate {
+            parameters["time"] = entitlementDate
             return parameters
         } else {
             return [:]